Allow users with corepathways access to be brokers
The corepathways Google group is used as a trigger to set the role to
'Referrer', however this should be in addition to regular roles and not
exclusive as it was previous to this commit. This is because there is
likely to be a significant number of people with access to both systems.